chore: rename `Sidebar`
