--- packages/fuselage/src/components/SideBarV2/SideBarFooter/index.ts
@@ -1,2 +1,0 @@
-export * from './SideBarFooter';
-export * from './SideBarFooterContent';
