--- packages/fuselage/src/components/SideBarV2/SideBarMedia/index.ts
@@ -1,3 +1,0 @@
-export * from './SideBarMedia';
-export * from './SideBarMediaController';
-export * from './SideBarMediaTitle';
