--- packages/fuselage/src/components/SideBarV2/index.tsx
@@ -1,15 +1,0 @@
-export * from './SideBar';
-export * from './SideBarAccordion';
-export * from './SideBarAccordionItem';
-export * from './SideBarAction';
-export * from './SideBarActions';
-export * from './SideBarLink';
-export * from './SideBarItem';
-export * from './SideBarCollapseGroup';
-export * from './SideBarBanner';
-export * from './SideBarFooter';
-export * from './SideBarMedia';
-export * from './SideBarButtonGroup';
-export * from './SideBarSection';
-export * from './SideBarDivider';
-export * from './SideBarGroupTitle';
