+++ packages/fuselage/src/components/SidebarV2/Sidebar.styles.scss
