+++ packages/fuselage/src/components/SidebarV2/SidebarAccordionItem.tsx
