+++ packages/fuselage/src/components/SidebarV2/SidebarFooter/SidebarFooter.styles.scss
