+++ packages/fuselage/src/components/SidebarV2/SidebarFooter/index.ts
@@ -1,0 +1,2 @@
+export * from './SidebarFooter';
+export * from './SidebarFooterContent';
