+++ packages/fuselage/src/components/SidebarV2/SidebarGroupTitle.tsx
