+++ packages/fuselage/src/components/SidebarV2/SidebarItem/SidebarItem.styles.scss
