+++ packages/fuselage/src/components/SidebarV2/SidebarItem/SidebarItemAvatarWrapper.tsx
