+++ packages/fuselage/src/components/SidebarV2/SidebarItem/SidebarItemContent.tsx
