+++ packages/fuselage/src/components/SidebarV2/SidebarItem/SidebarItemIcon.tsx
