+++ packages/fuselage/src/components/SidebarV2/SidebarMedia/SidebarMedia.stories.tsx
