+++ packages/fuselage/src/components/SidebarV2/SidebarMedia/index.ts
@@ -1,0 +1,3 @@
+export * from './SidebarMedia';
+export * from './SidebarMediaController';
+export * from './SidebarMediaTitle';
