+++ packages/fuselage/src/components/SidebarV2/hooks/useCollapse.tsx
