+++ packages/fuselage/src/components/SidebarV2/index.tsx
@@ -1,0 +1,15 @@
+export * from './Sidebar';
+export * from './SidebarAccordion';
+export * from './SidebarAccordionItem';
+export * from './SidebarAction';
+export * from './SidebarActions';
+export * from './SidebarLink';
+export * from './SidebarItem';
+export * from './SidebarCollapseGroup';
+export * from './SidebarBanner';
+export * from './SidebarFooter';
+export * from './SidebarMedia';
+export * from './SidebarButtonGroup';
+export * from './SidebarSection';
+export * from './SidebarDivider';
+export * from './SidebarGroupTitle';
